+++ lib/completion.dart
@@ -5,9 +5,18 @@ import 'src/try_completion.dart';
 
 export 'src/generate.dart';
 
-ArgResults tryArgsCompletion(List<String> mainArgs, ArgParser parser) {
-  tryCompletion(mainArgs, (List<String> args, String compLine, int compPoint) {
-    return getArgsCompletions(parser, args, compLine, compPoint);
-  });
+ArgResults tryArgsCompletion(
+    List<String> mainArgs,
+    ArgParser parser,
+    {@Deprecated('Useful for testing, but do not released with this set.')
+        logFile}) {
+  tryCompletion(
+    mainArgs,
+    (List<String> args, String compLine, int compPoint) {
+      return getArgsCompletions(parser, args, compLine, compPoint);
+    },
+    // ignore: deprecated_member_use
+    logFile: logFile,
+  );
   return parser.parse(mainArgs);
 }